requirements.txt
alembic==1.12.0
aniso8601==9.0.1
asttokens==2.2.1
attrs==21.2.0
backcall==0.2.0
blinker==1.6.2
cachelib==0.10.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6


cors==1.0.1
cryptography==3.4.8


decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0


email-validator==2.0.0.post2

exceptiongroup==1.1.3
executing==1.2.0
Faker==18.5.0
filelock==3.12.2
Flask==1.1.2
Flask-Cors==4.0.0
Flask-Login==0.4.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF<=1.1.1
fonttools<=4.38.0
future==0.18.3
gevent<=22.10.2
greenlet==2.0.2
gunicorn==21.2.0
httpie==3.2.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
ipdb==0.13.13
ipython<=7.34.0
itsdangerous==1.1.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==2.10.3
jsonschema==3.2.0
keyring==23.5.0

launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.2.4
Markdown==2.2.0
markdown-it-py==2.2.0
MarkupSafe==1.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.2

mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
openpyxl==3.1.2
packaging==23.1

parso==0.8.3

pexpect==4.8.0
pickleshare==0.7.5

pipenv==2023.7.23
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39

ptyprocess==0.7.0
pure-eval==0.2.2
psycopg2==2.9.9

Pygments==2.16.1

PyJWT==2.3.0
pyparsing==2.4.7

PySocks==1.7.1
pysqlite3==0.5.1
pytest==7.4.0


python-dateutil==2.8.2
python-dotenv==1.0.0

pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.6.0

SecretStorage==3.3.1
six==1.16.0

SQLAlchemy==1.4.25
SQLAlchemy-serializer==1.4.1
stack-data==0.6.2

tldextract==3.6.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1

urllib3==2.0.5

virtualenv==20.24.3
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.6
wheel<=0.37.1
Werkzeug==1.0.1
WTForms==3.0.1
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
